refactor, imporve
